swipeTracker: Move distance from individual gestures to SwipeTracker
We'll unify the event handling paths of the SwipeTracker to share a
single path for the pan gesture and for our self-implemented touchpad
ones. To make this easier, a few things need to be shuffled around.

Start by moving the "distance" that touchpad gestures add to their
"update" and "end" emissions to the SwipeTracker itself, this makes
their signals a bit more similar to the PanGesture signals.