overview: Add another workspacesView gesture up to the overview
We'll soon merge the workspaceAnimation gesture that switches
workspaces during the session with the workspacesView gesture inside the
overview. This makes it necessary that the workspacesView gesture is
added to the stage. Also it needs to be possible to show the overview
when the workspacesView gesture starts, so it makes sense to add
another horizontal gesture to the Overview class that controls the
workspacesView adjustment.