workspacesView: Change swipe distance when workspace size changes
The distance needed for the horizontal (workspaces) swipe tracker to
move by one full workspace changes with every update of the vertical
gesture, because every update of the vertical gesture changes the size
of the workspaces.

With our new two dimensional gesture it will be possible to do both
gestures at the same time, so take this size-change into account and
update the distance of the currently active workspaces swipe tracker
when the size of a workspace changed.