windowPreview: Don't report overlay in chrome sizes when its disabled
With the next commit we'll permanently hide the overlay when gnome-shell
is in phone mode. In this case we'll want the window previews to be
perfectly centered, without adjusting for things like the close button,
so don't report the overlay sizes in chromeWidths() and friends when
overlays are hidden.

(IN OTHER BRANCH) am window preview