appDisplay: Disallow dragging running apps from dash via _canAccept()
This is the preferred solution for a clean DND setup: Check if the app
is part of the running apps in the Dash, and then return false in
canAccept() for those icons.

Sadly we can't import the Dash here properly because that would be a
circular import, which apparently confuses gjs, so rather move the DashIcon
class into AppDisplay.