panel: Use different items when on mobile
On mobile we don't have that much horizontal screen space to show lots
of things in the panel, so hide most things there and only show the most
important things.

Slightly annoying but we'll have to keep showing the keyboard indicator
since that class is setting up crucial things for the OSK to work.