overview: Disallow window picker state in mobile mode
The current design direction for mobile is that it doesn't make sense to
show the window picker on mobile, given that only one window can be on
a workspace at a time.

Sadly implementing this is fairly ugly, we do it by pretending the
WINDOW_PICKER state is a substate of the APP_GRID transition and then
splattering a few is_phone checks all over the place to avoid weird
in-between states during the transition.