appDisplay: Extend app grid up to horizontal screen edges on phones
On desktops we add some margin to the left and right side of the appGrid
because most screens are too large for it to look nice. Phones have
small screens though, we don't want to waste precious pixel real estate
there, so extend the appGrid up to the whole horizontal size (and
hide the hint to show the next page) there.