style/search: Improve look of results on mobile layouts
Make the layout of per-app results vertical instead of horizontal, make the
background semi-transparent to allow for overview wallpaper, and ensure the
font is white-ish instead of gray.

With the wallpaper behind, the background can be quite noisy, so add a bit
of text-shadow to ensure readability. We use the same shadow as on app item
labels.

Also make the spacing between the app search results at the top a lot
smaller.