cldr2json: Autogenerate mobile OSK layouts for phones
Add some quite specific code in order to autogenerate layouts for mobile OSK
layouts. This is supposed to bootstrap as many useful mobile layouts as
possible, so that we can then iterate and update the specific layouts from
there. This is *not* meant to be adapted and to regenerate all mobile
layouts again at some point later, instead after the bootstraping the
layouts should be improved by hand.