panel: Add a pull-down gesture to show and hide quicksettings
Nothing fancy, use translations to move the boxPointer container driven by a
pan gesture.

Since we only want the popupMenu to take its grab after the gesture is fully
done, set the "modal" property of the menu to false until the gesture has
finished. In case the gesture didn't finish, reset the property back to
true, so that the next one calling open() won't be surprised by a non-modal
behavior.