unlockDialog: Don't show notifications in prompt mode
On mobile we'll want to show a keypad in prompt mode. That means there'll be
no more space to show the notifications in that mode. So show the only in
the clock mode, and not in the other stack page.

While at it, respect the expand property of the clock in the layout manager
so that things look alright.