popupMenu: Allow opening without a grab
There's a few cases where we want to open a popup menu without taking the
grab immediately. That's because the menu is opened by a gesture and we
don't know yet whether that gesture will succeed. If it won't succeed, we
don't want to take a grab.

The solution for this is to take the grab later, so we need to decouple
showing the popupMenu from actually taking its grab.

To do this, we introduce a new "modal" property on the PopupMenu base
class which defaults to true. It can be set to false before the open() in
order to avoid taking the grab, and it can be set to true later while
the menu is already open in order to take the grab later. Going from modal
to !modal while the menu is already open is not possible.