sensorDaemon: add calls monitoring
This is really not how it should work conceptually: We shouldn't be talking
to GNOME Calls, but rather Calls should be talking to us, and tell us to
enable proximity monitoring.

Anyway, for now do the same thing that Phosh does and talk to GNOME Calls
via DBus to be notified about any ongoing calls.