keyboard: Introduce a custom ClutterGesture for the OSK
Quite the big one, this one introduces a custom gesture for the OSK that
allows for fancier input processing than what's possible so far. Instead
listening for "clicked" signals on the individual buttons that make up the
OSK keys, have a single gesture that spans across the whole OSK.

The new thing looks up the key that has been pressed by checking in its own
representation of the key-grid instead of doing picking. This allows for
more dynamic "hit boxes" of the keyboard keys, interpreting touches in
between two keys or in the empty space at the keyboard edge as key presses
just like most other platforms do.