swipeTracker: Remove touchpad event handling code
This is now handled by the pan gesture itself, we don't have to care in the
swipeTracker anymore.

Thanks to the pan gesture, the events are now also handled solely on the
actor, no event handlers are attached to the stage. This allows removing
a hack that was needed to disable the swipeTracker of the emojiPager while
its actor was hidden (see a0352ba3).