keyboard: Make the keyboardBox extend the whole stage
With the next commit, we'll port the OSK to the new ClutterGesture system.
To implement the closing of the OSK (subkeys and popupMenu) overlays there,
we'll use a "cover actor" that spans the whole uiGroup.

To make this work properly, the cover actor needs to be in between the
overlay and the OSK itself. This is not possible right now because the
keyboardBox is only the size of the keyboard itself, so our cover actor
could not be a larger than that box. To allow a cover actor that
extends over the whole stage, we need to make the keyboardBox itself
extend over the whole stage, so let's do that.