sessionMode: Formally allow notifications on lockscreen
We already show notifications on the lockscreen, just using a different
widget than our usual ones. We're going to start reusing those for the
lockscreen though, so let's formally allow notifications in the unlockDialog
session mode.

We'll rather make sure that no sensitive notification content leaks by
making the NotificationSection class aware that it's running on the
lockscreen.