screenShield: Refactor slide-up animation on deactivate()
Right now we start two conflicting animations to slide up
this._lockDialogGroup inside _continueDeactivate(): The first one inside
_hideLockScreen(), and then another one at the end of _continueDeactivate().

Clean this up by moving the contents of _hideLockScreen() into
_continueDeactivate() so that we can call _completeDeactivate() from the
onComplete() callback of the actual slide-up animation.

Note that this changes the this._setActive(false) call when we're in greeter
mode to happen *after* the animation, not before, this should be correct
though and is just like we do it in the !greeter mode.