schedulingManager: Improve animation smoothness
Utilization clamping hints the Linux scheduler about the workload of a
given process. Some ARM CPUs have a big.LITTLE architecture to balance
between CPU performance and power efficiency.

The scheduling manager uses minimal utilization clamping to hint the
scheduler that a given process should be scheduled on the performant
cores instead of the efficient ones. This improves UI smoothness.

The hinting is only active for the focused application and for the
shell process when the user is not idle.