keyboard: Use KEY_SIZE for grid layout of emoji keyboard too
In theory the emoji keyboard could get away without multiplying by KEY_SIZE
for its grid layout (because the pager has a height of 3 and width of 1,
while the bottom row has a height of 1, and width of 1, so no rows or cols
in the grid are smaller than 1).

We use the KEY_SIZE for the keys in the normal, non-emoji OSK though, and
there it triggers a weird bug: ClutterGridLayout allocates the first row
slightly taller, second row less tall ... and the last row the least tall.
It's a subtle, hardly noticable difference, but it does become noticable
when switching to the emoji page, because here the last row has the correct
height, and that is taller than the last row in the non-emoji page.

So work around this weird bug in the grid layout, and make those two row
heights match, by multiplying the emoji grid layout with KEY_SIZE, too.