keyboard: Handle small screens in AspectContainer
The AspectContainer tries to apply an aspect ratio to the natural
width/height, using the given forWidth/Height. If the available screen size
is too small for the aspect ratio to be fulfilled though, this makes sure
that the natural size can still be used.

I'm not sure if this is 100% correct though..