keyboard: Add a double click gesture on the bottom bar to trigger the OSK
Lots of clients (looking at you chromium) aren't supporting wl-text-input v3
yet, and therefore the OSK will not automatically show up in those clients.

In mobile-shell the swipe-up gesture from the bottom screen edge is already
used, and not having any way to show the OSK is bad, so add a double tap
gesture on the bottom screen edge to allow for showing the OSK manually.