keyboard: Always use keyval instead of wl-text-input for the backspace key
The newer method based on wl-text-inputs delete_surrounding() method works
fairly well now with gtk4, but it has significant problems with gtk3,
firefox and other clients.

In mobile-shell the OSK is the only method to input text, so default to the
old, keyval-based, method to make sure it always works.