st/viewport: HACK: work around some bug with size calcualtion
There's a very nasty bug somewhere that appears to be some kind of rounding
error that is scaling or wayland related (it only happens on the 3x wayland
session on the op6, but not in nested session).

This bug causes the min_height reported by ClutterLayutManager to be 1px
larger than the avail height that StViewport get from the layout machinery,
and that in turn means the notifications are scrollable/fade is enabled even
though there is enough space.

Since the error is just 1px, subtract that pixel for now..