powerManager: Add a power manager module to handle power states and events
This is mostly trying to re-implement the features exactly as gsd-power was
doing things.

Having all this in gnome-shell allows us to do a lot of things in better
integrated ways, eg. showing an animation when turning of the screen, or
locking the screen on suspend without race conditions.

There's a bunch of things that gsd-power handled that we no longer handle
here, that's mainly:

- lid-open/close events from UPower: These are handled by systemd-logind
these days, no need to do that ourselves
- inhibiting lid-switch suspend when there's more than one monitor: This
also seems to be handled by logind
- dimming more agressively when either the power-saver profile is enabled,
or when the battery is low (in case there's no power-profiles-daemon)
- handling UPower notifications (UPS in-use, critical battery, fully charged
etc) and enabling/disabling the power-saver profile is still left to
gsd-power-manager, as that is fairly self-contained and can remain
- auto-backlight adjusting/dimming for screen and setting keyboard backlight
brightness remains remains in gsd-power for now and will move to a separate
daemon in the future