keyboard: Add press-feedback and move keys up while pressed
On mobile it's important to have a bit more feedback than just the
background-color changes to keys, because the finger is usually covering the
whole key, making the color change moot.

Instead, use translation-y to crudely move the key up while pressed. In
a few tests, this works really well and allows for faster typing.