keyboard: Make latching gesture a double click and change levels on press
In general double clicks are quicker to execute than long presses, and this
is the way both Android and iOS OSKs are doing it these days.

There's another reason for this though: We want to change the latching keys
to react on press, not release (it feels a lot snappier, and again both
Android and iOS do it), and with that the layout will change on press. This
will make it very painful to keep using a long-press gesture for latching,
because the KeyContainer that the gesture lives on gets unmapped when
switching the level, cancelling the long-press gesture. Moving to a double
click gesture is kind of a nice trick to avoid having to deal with that.