unlockDialog: Always move focus into authPrompt when it doesn't have focus
Sometimes on the lockscreen, key focus is on the UnlockDialog as expected,
but not on the authPrompt entry (even though the authPrompt is the active
page of the unlockDialog). This leads to key presses getting swallowed and
not ending up in the password entry.

Fix that by explicitly checking whether focus is inside the authPrompt, and
if it isn't, call AuthPrompt.addCharacter(), which will move focus to the
password entry.