authPrompt: Set status back to VERIFYING after failure
When starting to verify, we set the status to VERIFYING, then when
submitting our answer, we move to VERIFICATION_IN_PROGRESS, and then when it
fails and we can't retry, we set status to VERIFICATION_FAILED, but in case
we *can* retry, we leave it set to VERIFICATION_IN_PROGRESS. Since the next
call is to _activeNext() again (which sets status to
VERIFICATION_IN_PROGRESS), it makes sense to set the state back to
VERIFYING here.

This will help with the next commit.