style/keyboard: Make key font-size smaller in mobile + horizontal layouts
On mobile horizontal layouts we're very constrained in vertical space for
the OSK, because phone monitors are small. To allow fitting all the OSK rows
on the screen, use a smaller font size for the labels of the OSK keys.

Possibly a Cantarell issue: When increasing the font size just one pt more
from what we have now, apparently the text requests a lot more size, so the
keyboard gets a ton larger even though the text easily fits within the keys.