workspace: Update wallpaper on monitor changes
Since the last commit (ab1b6252ad (overview: Don't hide on monitor
changes)), the overview now remains opened on monitor changes.

This exposed a small bug: The wallpapers that we show behind the individual
workspace previews in the overview should also be updated when the monitor
changes, let's do that.