keyboard: Allow using extra small font-size for labels of certain keys
Sometimes keys (for example ﷼  ) in the Persian layout are slightly too
large and won't fit into the default size of the keys of the OSK. This will
cause all keys of the OSK to become larger (because the grid-layout is
homogeneous), and therefore the OSK to suddenly change size when switching
to that layout.

We don't want those sudden size changes, so for those special larger keys,
allow setting a smallerFontSize property in the layout definition, and then
use a font size of 1pt less than the default.