overview: Don't hide on monitor changes
The overview is laid out mostly through the normal layout machinery these
days, so it responds just fine to size-changes of the stage. Hiding it and
papering over the relayout issues should no longer be necessary, remove it.

This is especially annoying on phones, because we're also closing the
overview on screen rotations.