osk-layouts: Generate mobile and mobile-extended layouts for all languages
Make use of the updated cldr2json script in order to auto-generate mobile
OSK layouts for all the languages. This is probably not correct for every
language, as I mainly tested the english and german mobile layouts, but it's
a start for mobile layouts at least.

Improvements to those layouts will hopefully be a crowdsourced effort
later.