FIXME: keyboard: Allow setting a right offset for keys in OSK layouts
Having a right offset property for keys is nice in order to be able to keep
a distance from the right monitor edge (where a leftOffset won't help us
anymore). We'll make use of it in some of the mobile layouts.

FIXME: the commit message doesn't reflect everything this does, also this is
a pretty ugly approach with the findMultiple stuff...