keyboard: Show/hide currentLayout rather than currentPage to show emoji OSK
The this._currentLayout container contains actors for all the pages, and is
a sibling of the emoji panel (this._emojiSelection). Therefore it makes
sense to switch between the two by hiding one of them and showing the other,
rather than to make it more complicated by hiding the current page.