keyboard: Move the OSK gesture from individual pages to the layout container
We already have the OSK gesture live on the page rather than the individual
keys in order to make several fancy fast-typing tricks possible. There's a
few reasons to have the gesture even be aware of pages of the OSK, eg. to
allow a gesture like "press and hold shift -> drag finger to uppercase key
-> release and go back to lowercase layer".

Another compelling reason to have the gesture on the layout container is so
that the gesture doesn't get cancelled when switching pages. Right now when
pressing eg. shift to switch pages, the page switch happens on press, which
hides the current page and therefore cancels the gesture. The gesture only
recognizes on release though (for good reason, eg. you want to be able to
swipe down the OSK even when the swipe begins on shift key), so that means
the gesture never recognizes and instead cancels. This in turn triggers
mutter to forward the event to clients, doing unexpected things like
hiding popups etc, and we don't want that.

So move the gesture to exist on the "layout" container that includes all
the pages rather than on the individual pages.

Closes: https://gitlab.gnome.org/verdre/gnome-shell-mobile/-/issues/65