keyboard: Move popovers up while the OSK is visible
We already resize windows when the OSK is visible, and we don't do that
using the work-area or even MetaConstraints (that's because using the
work-area would resize all the windows, not just the currently focused one,
and that's a performance issue. We don't use constraints because there's
not a lot to gain from that, and it would cause extra trouble with oversized
windows and leave us with less control in the end).

That approach means that popovers don't get moved when the OSK is shown, so
those might still end up obscured by the OSK. Deal with that by introducing
extra logic to temporarily move up popovers as well. This works a lot better
than using the work-area or MetaConstraints, as clients apparently really
don't like their popovers to ever be resized/flipped/moved after showing
them once, and might just end up closing them again.

Ideally all of this would use insets eventually, so this can be seen as
somewhat temporary anyway.

Closes: https://gitlab.gnome.org/verdre/gnome-shell-mobile/-/issues/66