layout: Return EVENT_PROPAGATE in hot corner event handling
Clutter is enforcing a return value of EVENT_PROPAGATE for crossing events
since a while now, logging a warning and overriding if EVENT_STOP is
returned.

Clearly the hot corner still works, so it seems like this comment can be
removed and the return value changed, fixing a warning on the logs.